improved script

+++ run_tests.sh
@@ -1,21 +1,35 @@
 #!/usr/bin/env bash
+set -e
 
 MAX_COMPLEXITY="D"      # can be from "A" to "F" where "A" is the lowest
+SKIP_DIRS="env,tests"
 
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+
+echo -e "Activating env..."
 source ${DIR}/env/bin/activate
+
+echo -e "Export environ vars..."
 export PYTHONPATH=${PYTHONPATH}:${DIR}
+
+echo -e "Run tests..."
 nosetests . --config=nose.cfg
 
-radon cc -i "env,tests" --order=SCORE -s --xml . > complexity.xml
+echo -e "Compute CC..."
+radon cc -i ${SKIP_DIRS} --order=SCORE -s --xml . > complexity.xml
 
-radon cc -i "env,tests" --order=ALPHA -s -n=${MAX_COMPLEXITY} . 1>&2
+echo -e "Check CC"
+radon cc -i ${SKIP_DIRS} --order=ALPHA -s -n=${MAX_COMPLEXITY} . 1>&2
 
-flake8 . --output-file flake8.txt
+if [[ $(radon cc -i ${SKIP_DIRS} --order=ALPHA -s -n=${MAX_COMPLEXITY} . ) ]]; then
+    exit 1
+fi
 
+echo -e "Run flake8..."
+flake8 . --output-file flake8.txt || true       # continue execution if failed
+echo -e "Create flake8 xml file"
 flake8_junit flake8.txt flake8_junit.xml
-
-echo "\n"
-flake8 --statistics --count -qq 1>&2
-
 rm flake8.txt
+
+echo -e "\nCheck flake8..."
+flake8 --statistics --count -qq 1>&2
